Entrega 23-07

+++ app/Http/Controllers/AdminControllers/EvaluationsController.php
@@ -122,8 +122,11 @@ class EvaluationsController extends Controller
     public function destroy($id)
     {
         $evaluation = Evaluation::find($id);
-        $evaluation->delete();
-        return redirect()->route('evaluations.index');
+        // dd($evaluation);
+        if($evaluation != null){
+            $evaluation->delete();
+        }
+        return back();
     }
 
     public function dropImgAttachments($evaluation){
